# xhs_backend/api_docs/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "小红书评论维护系统",
    "description": "小红书评论数据的管理和维护系统",
    "version": "1.0.0"
  },
  "paths": {
    "/api/users/register": {
      "post": {
        "tags": [
          "用户",
          "认证"
        ],
        "summary": "Register",
        "description": "注册新用户\n\nArgs:\n    user_in: 用户注册信息\n    \nReturns:\n    访问令牌",
        "operationId": "register_api_users_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "用户",
          "认证"
        ],
        "summary": "Login",
        "description": "用户登录\n\nArgs:\n    user_in: 登录信息，包含用户名、密码和OTP代码\n    \nReturns:\n    访问令牌",
        "operationId": "login_api_users_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/otp-qrcode": {
      "get": {
        "tags": [
          "用户",
          "用户"
        ],
        "summary": "Get Otp Qrcode",
        "description": "获取当前登录用户的OTP二维码\n\nArgs:\n    current_user: 当前用户名\n    \nReturns:\n    OTP二维码图像",
        "operationId": "get_otp_qrcode_api_users_otp_qrcode_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/users/me": {
      "get": {
        "tags": [
          "用户",
          "用户"
        ],
        "summary": "Get Current User Info",
        "description": "获取当前登录用户信息\n\nArgs:\n    current_user: 当前用户名\n    \nReturns:\n    用户信息",
        "operationId": "get_current_user_info_api_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/users/info/{user_id}": {
      "get": {
        "tags": [
          "用户",
          "小红书用户"
        ],
        "summary": "Get Xhs User Info",
        "description": "获取小红书用户的信息\n\nArgs:\n    user_id: 小红书用户ID\n    current_user: 当前认证用户名\n    \nReturns:\n    用户信息或404",
        "operationId": "get_xhs_user_info_api_users_info__user_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            },
            "name": "user_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/users/info": {
      "get": {
        "tags": [
          "用户",
          "小红书用户"
        ],
        "summary": "Get Multiple Xhs User Info",
        "description": "批量获取多个小红书用户的信息\n\nArgs:\n    user_ids: 逗号分隔的小红书用户ID列表\n    current_user: 当前认证用户名\n    \nReturns:\n    用户信息映射",
        "operationId": "get_multiple_xhs_user_info_api_users_info_get",
        "parameters": [
          {
            "description": "逗号分隔的用户ID列表",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Ids",
              "description": "逗号分隔的用户ID列表"
            },
            "name": "user_ids",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/users/info/list": {
      "get": {
        "tags": [
          "用户",
          "小红书用户"
        ],
        "summary": "List Xhs Users",
        "description": "分页获取小红书用户信息列表\n\nArgs:\n    pagination: 分页参数\n    current_user: 当前认证用户名\n\nReturns:\n    分页的用户信息列表",
        "operationId": "list_xhs_users_api_users_info_list_get",
        "parameters": [
          {
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1.0,
              "title": "Page",
              "description": "页码",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "description": "每页项目数",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100.0,
              "minimum": 1.0,
              "title": "Page Size",
              "description": "每页项目数",
              "default": 20
            },
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
    "/api/register": {
      "post": {
        "tags": [
          "认证",
          "认证"
        ],
        "summary": "Register",
        "description": "注册新用户\n\nArgs:\n    user_in: 用户注册信息\n    \nReturns:\n    访问令牌",
        "operationId": "register_api_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/login": {
      "post": {
        "tags": [
          "认证",
          "认证"
        ],
        "summary": "Login",
        "description": "用户登录\n\nArgs:\n    user_in: 登录信息，包含用户名、密码和OTP代码\n    \nReturns:\n    访问令牌",
        "operationId": "login_api_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/otp-qrcode": {
      "get": {
        "tags": [
          "认证",
          "用户"
        ],
        "summary": "Get Otp Qrcode",
        "description": "获取当前登录用户的OTP二维码\n\nArgs:\n    current_user: 当前用户名\n    \nReturns:\n    OTP二维码图像",
        "operationId": "get_otp_qrcode_api_otp_qrcode_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/me": {
      "get": {
        "tags": [
          "认证",
          "用户"
        ],
        "summary": "Get Current User Info",
        "description": "获取当前登录用户信息\n\nArgs:\n    current_user: 当前用户名\n    \nReturns:\n    用户信息",
        "operationId": "get_current_user_info_api_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/info/{user_id}": {
      "get": {
        "tags": [
          "认证",
          "小红书用户"
        ],
        "summary": "Get Xhs User Info",
        "description": "获取小红书用户的信息\n\nArgs:\n    user_id: 小红书用户ID\n    current_user: 当前认证用户名\n    \nReturns:\n    用户信息或404",
        "operationId": "get_xhs_user_info_api_info__user_id__get",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            },
            "name": "user_id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/info": {
      "get": {
        "tags": [
          "认证",
          "小红书用户"
        ],
        "summary": "Get Multiple Xhs User Info",
        "description": "批量获取多个小红书用户的信息\n\nArgs:\n    user_ids: 逗号分隔的小红书用户ID列表\n    current_user: 当前认证用户名\n    \nReturns:\n    用户信息映射",
        "operationId": "get_multiple_xhs_user_info_api_info_get",
        "parameters": [
          {
            "description": "逗号分隔的用户ID列表",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Ids",
              "description": "逗号分隔的用户ID列表"
            },
            "name": "user_ids",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/info/list": {
      "get": {
        "tags": [
          "认证",
          "小红书用户"
        ],
        "summary": "List Xhs Users",
        "description": "分页获取小红书用户信息列表\n\nArgs:\n    pagination: 分页参数\n    current_user: 当前认证用户名\n\nReturns:\n    分页的用户信息列表",
        "operationId": "list_xhs_users_api_info_list_get",
        "parameters": [
          {
            "description": "页码",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1.0,
              "title": "Page",
              "description": "页码",
              "default": 1
            },
            "name": "page",
            "in": "query"
          },
          {
            "description": "每页项目数",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100.0,
              "minimum": 1.0,
              "title": "Page Size",
              "description": "每页项目数",
              "default": 20
            },
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TokenResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "TokenResponse"
      },
      "UserInLogin": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          },
          "otp_code": {
            "type": "string",
            "title": "Otp Code"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password",
          "otp_code"
        ],
        "title": "UserInLogin"
      },
      "UserInRegister": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "UserInRegister"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/login"
          }
        }
      }
    }
  }
}